+++ src/features/Login/services/oauth.ts
@@ -1,0 +1,4 @@
+const REST_API_KEY = import.meta.env.VITE_APP_REST_API_KEY;
+const REDIRECT_URI = import.meta.env.VITE_APP_REDIRECT_URL;
+
+export const KAKAO_AUTH_URL = `https://kauth.kakao.com/oauth/authorize?client_id=${REST_API_KEY}&redirect_uri=${REDIRECT_URI}&response_type=code`;
